+++ .babelrc
@@ -1,13 +1,13 @@
 {
-    "presets": [
-        [
-            "@babel/preset-env",
-            {
-                "targets": {
-                    "node": "6.10",
-                    "esmodules": true
-                }
-            }
-        ]
+  "presets": [
+    [
+      "@babel/preset-env",
+      {
+        "targets": {
+          "node": "6.10",
+          "esmodules": true
+        }
+      }
     ]
+  ]
 }
